feat(providers): complete package migration, old module cleanup, and synapse integration (#299)
Phase 5: Finalize multi-LLM provider support.

- Remove deprecated memory/spector-embed-api and memory/spector-embed-ollama modules from parent POM and filesystem
- Update package imports in 25+ Java files across all modules from com.spectrayan.spector.embed.ollama to com.spectrayan.spector.provider.ollama
- Update spector.yml.example and spector-defaults.yml to document the new spector.provider.* configuration namespace
- Add compatibility methods (active name, health indicators, allHealth, and alias names) to ProviderRegistry interface and DefaultProviderRegistry to maintain full backward compatibility with Synapse management endpoints
- Remove duplicate provider classes under spector-synapse provider package and update DelegatingTextGenerationProvider, ProviderController, and EmbeddingProviderConfig to leverage Core's new ProviderRegistry and DefaultProviderRegistry beans
- Resolve float-to-Double boxing issue in LangChain4jGenerationAdapter
- Verify with clean reactor build and test execution (BUILD SUCCESS)

+++ memory/spector-providers/src/main/java/com/spectrayan/spector/provider/langchain4j/LangChain4jGenerationAdapter.java
@@ -79,12 +79,12 @@ public class LangChain4jGenerationAdapter implements TextGenerationProvider {
 
         // Build per-request parameters from GenerationOptions
         var paramsBuilder = ChatRequestParameters.builder();
-        paramsBuilder.temperature(options.temperature());
+        paramsBuilder.temperature((double) options.temperature());
 
         if (options.maxTokens() > 0) {
             paramsBuilder.maxOutputTokens(options.maxTokens());
         }
-        paramsBuilder.topP(options.topP());
+        paramsBuilder.topP((double) options.topP());
 
         if (options.stopSequences() != null && options.stopSequences().length > 0) {
             paramsBuilder.stopSequences(List.of(options.stopSequences()));
